MAINT: slightly refactored config module
